refactor(codebase): implement changes to the codebase based on review [Finishes #140492107]
